extract so we can test
