+++ src/lib/utils/track/index.tsx
@@ -1,9 +1,10 @@
-import { ActionType, Screen } from "@artsy/cohesion"
+import { Screen } from "@artsy/cohesion"
 import _track, { Track as _Track, TrackingInfo } from "react-tracking"
 
 // The schema definition for analytics tracking lives inside `./schema`, not here.
 import React from "react"
 import * as Schema from "./schema"
+import { postEventToProviders, TrackingProvider, _addTrackingProvider } from "./providers"
 export { Schema }
 
 /**
@@ -225,42 +226,6 @@ export function screenTrack<P>(trackingInfo: TrackingInfo<Schema.PageView, P, nu
  *
  */
 
-interface CohesionAction {
-  // TODO: This can be removed once cohesion provides a global `Action` type.
-  action: string
-}
-interface CohesionScreen {
-  // TODO: This can be removed once cohesion provides a global `Action` type.
-  action: ActionType.screen
-  context_screen: string
-}
-export const isCohesionScreen = (info: CohesionAction | CohesionScreen): info is CohesionScreen =>
-  info.action === ActionType.screen
-
-interface LegacyNameAction {
-  // TODO: This can be removed once we remove these uses. Currently there is one in `WorksForYou` and one in `ArtistRail`.
-  name: string
-}
-
-type InfoType = Schema.PageView | Schema.Entity | CohesionAction | CohesionScreen | LegacyNameAction
-
-export interface TrackingProvider {
-  identify?: (userId: string | null, traits?: { [key: string]: any }) => void
-  postEvent: (info: InfoType) => void
-}
-
-const providers: { [name: string]: TrackingProvider } = {}
-
 export const addTrackingProvider = (name: string, provider: TrackingProvider) => {
-  providers[name] = provider
-}
-
-export const postEventToProviders = (info: any) => {
-  Object.values(providers).forEach((provider) => {
-    provider.postEvent(info)
-
-    if (__DEV__) {
-      console.log("[Event tracked]", JSON.stringify(info, null, 2))
-    }
-  })
+  _addTrackingProvider(name, provider)
 }
